// resources/assets/src/app/models/api/response/CategoriesReponse.ts
export class CategoryReponse {
    id: number;
    name: string;
    display_name: string;
    description: string;
    created_at: string;
    updated_at: string;
    constructor() {

    }
}

export class HolidayReponse {
    id: number;
    start_date: string;
    end_date: string;
    description: string;
    created_at: string;
    updated_at: string;
    constructor() {

    }
}

export class CategoryOtherReponse {
    id: number;
    name: string;
    type: string;
    description: string;
    created_at: string;
    updated_at: string;
    constructor() {

    }
}